Ribbon - Add convex hull
